requirements.txt
streamlit
browser-use
playwright
google-generativeai
python-dotenv
requests